Prevent manual file nesting across directories.
The Nest-Item menu is now disabled if any of the selected files are
in a different directory. Whilst it was possible to nest files in
different directories and it looks like it works in the Solution
window on reloading the project in the IDE would not show the files
as nested since they need to be in the same directory.